Renamed GoldenEye.Core.Dapper into GoldenEye.Dapper Fixed namespaces, updated packages to latest versions, test projects to .NET 5
