+++ src/Dapper/Dapper/Readme.md
@@ -1,11 +1,11 @@
-﻿# GoldenEye.Backend.Core.Dapper
+﻿# GoldenEye.Dapper
 [![Stories in Ready](https://badge.waffle.io/oskardudycz/GoldenEye.png?label=ready&title=Ready)](https://waffle.io/oskardudycz/GoldenEye)
 [![Build status](https://ci.appveyor.com/api/projects/status/1mtm4h33cvur6kob?svg=true)](https://ci.appveyor.com/project/oskardudycz/goldeneye-core)
-[![NuGet version](https://badge.fury.io/nu/GoldenEye.Backend.Core.Dapper.svg)](https://badge.fury.io/nu/GoldenEye.Backend.Core.Dapper)
+[![NuGet version](https://badge.fury.io/nu/GoldenEye.Dapper.svg)](https://badge.fury.io/nu/GoldenEye.Dapper)
 
-What is GoldenEye.Backend.Core.Dapper?
+What is GoldenEye.Dapper?
 --------------------------------
-GoldenEye.Backend.Core.Dapper is a library that brings you abstractions and implementations for common backend topics. It is written in .NET Core. It provides set of base and bootstrap classes that helps you to reduce boilerplate code and help you focus on writing business code.
+GoldenEye.Dapper is a library that brings you abstractions and implementations for common backend topics. It is written in .NET Core. It provides set of base and bootstrap classes that helps you to reduce boilerplate code and help you focus on writing business code.
 
 What do I get?
 --------------------------------
@@ -15,9 +15,9 @@ Complete helpers and bootstrap for:
 
 How do I get started?
 --------------------------------
-You can create new project and add [nuget package](https://www.nuget.org/packages/GoldenEye.Backend.Core.Dapper):
+You can create new project and add [nuget package](https://www.nuget.org/packages/GoldenEye.Dapper):
 
-`dotnet add package GoldenEye.Backend.Core.Dapper`
+`dotnet add package GoldenEye.Dapper`
 
 I found an issue or I have a change request
 --------------------------------
