+++ src/GoldenEye.sln
@@ -39,8 +39,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Backend.Core.WebApi.Tests",
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoldenEye.WebApi.Template.SimpleDDD", "Templates\SimpleDDD\content\GoldenEye.WebApi.Template.SimpleDDD.csproj", "{592DD3EB-823E-4941-808A-7A42928748BD}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Backend.Core.Dapper", "Core\Backend.Core.Dapper\Backend.Core.Dapper.csproj", "{7D9EBFA0-F65B-4F87-824A-1F3F2D18BD0A}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Docker", "Docker", "{745EC139-4033-4F75-BC9C-64A417DC1229}"
 	ProjectSection(SolutionItems) = preProject
 		..\docker\docker-compose.yml = ..\docker\docker-compose.yml
@@ -61,10 +59,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SimpleDDD", "SimpleDDD", "{
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Dapper", "Dapper", "{E15E5EE4-FD58-4081-AA20-1B74C162D3E3}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Backend.Core.Dapper.Tests", "Core\Backend.Core.Dapper.Tests\Backend.Core.Dapper.Tests.csproj", "{326790CA-C54A-4DC3-A574-CC054AF56CC2}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Backend.Core.Dapper.Integration.Tests", "Core\Backend.Core.Dapper.Integration.Tests\Backend.Core.Dapper.Integration.Tests.csproj", "{C522605E-EBB8-4D37-AD4B-60A096D52D34}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DDD", "DDD", "{1A32473D-9490-4FF8-89E7-986DCCF4599C}"
 EndProject
@@ -91,6 +85,12 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core\Core.csproj", "{CD33583D-EB52-4D8A-8892-18FE4840CA06}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core.Tests", "Core\Core.Tests\Core.Tests.csproj", "{A84F8BA7-9E7D-4B49-B56B-F17B206B33B8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dapper", "Dapper\Dapper\Dapper.csproj", "{232C4AA8-172D-43F3-AD53-04306D70C128}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dapper.Tests", "Dapper\Dapper.Tests\Dapper.Tests.csproj", "{70A0FDC0-26E5-4480-AE9C-96C9C424AEC8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dapper.Integration.Tests", "Dapper\Dapper.Integration.Tests\Dapper.Integration.Tests.csproj", "{55AC76A5-1297-4529-8B29-5548EA581D22}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -130,22 +130,10 @@ Global
 		{592DD3EB-823E-4941-808A-7A42928748BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{592DD3EB-823E-4941-808A-7A42928748BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{592DD3EB-823E-4941-808A-7A42928748BD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7D9EBFA0-F65B-4F87-824A-1F3F2D18BD0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7D9EBFA0-F65B-4F87-824A-1F3F2D18BD0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7D9EBFA0-F65B-4F87-824A-1F3F2D18BD0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7D9EBFA0-F65B-4F87-824A-1F3F2D18BD0A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8E2E928B-E1D1-4AFA-AAE7-6586F9EC8170}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8E2E928B-E1D1-4AFA-AAE7-6586F9EC8170}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8E2E928B-E1D1-4AFA-AAE7-6586F9EC8170}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8E2E928B-E1D1-4AFA-AAE7-6586F9EC8170}.Release|Any CPU.Build.0 = Release|Any CPU
-		{326790CA-C54A-4DC3-A574-CC054AF56CC2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{326790CA-C54A-4DC3-A574-CC054AF56CC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{326790CA-C54A-4DC3-A574-CC054AF56CC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{326790CA-C54A-4DC3-A574-CC054AF56CC2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C522605E-EBB8-4D37-AD4B-60A096D52D34}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C522605E-EBB8-4D37-AD4B-60A096D52D34}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C522605E-EBB8-4D37-AD4B-60A096D52D34}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C522605E-EBB8-4D37-AD4B-60A096D52D34}.Release|Any CPU.Build.0 = Release|Any CPU
 		{40CCCF55-F3FB-4C13-8A1D-C737699C1CAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{40CCCF55-F3FB-4C13-8A1D-C737699C1CAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{40CCCF55-F3FB-4C13-8A1D-C737699C1CAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -182,6 +170,18 @@ Global
 		{A84F8BA7-9E7D-4B49-B56B-F17B206B33B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A84F8BA7-9E7D-4B49-B56B-F17B206B33B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A84F8BA7-9E7D-4B49-B56B-F17B206B33B8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{232C4AA8-172D-43F3-AD53-04306D70C128}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{232C4AA8-172D-43F3-AD53-04306D70C128}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{232C4AA8-172D-43F3-AD53-04306D70C128}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{232C4AA8-172D-43F3-AD53-04306D70C128}.Release|Any CPU.Build.0 = Release|Any CPU
+		{70A0FDC0-26E5-4480-AE9C-96C9C424AEC8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70A0FDC0-26E5-4480-AE9C-96C9C424AEC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70A0FDC0-26E5-4480-AE9C-96C9C424AEC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70A0FDC0-26E5-4480-AE9C-96C9C424AEC8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{55AC76A5-1297-4529-8B29-5548EA581D22}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{55AC76A5-1297-4529-8B29-5548EA581D22}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{55AC76A5-1297-4529-8B29-5548EA581D22}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{55AC76A5-1297-4529-8B29-5548EA581D22}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -196,12 +196,9 @@ Global
 		{E7DF2D12-586C-4FC6-AD53-082A1B5DF85B} = {23C64FA6-FA22-4A73-8360-D81FD44E7AE1}
 		{9BDEDD7C-A07A-42BD-85EB-2A7F29F10E1E} = {61870107-24D4-438B-B592-5A6868143AD0}
 		{592DD3EB-823E-4941-808A-7A42928748BD} = {54E79D60-4472-45EF-AA93-764C706453CC}
-		{7D9EBFA0-F65B-4F87-824A-1F3F2D18BD0A} = {E15E5EE4-FD58-4081-AA20-1B74C162D3E3}
 		{E9097CBA-BBAA-4E27-BA2E-1DF61CBFCB37} = {745EC139-4033-4F75-BC9C-64A417DC1229}
 		{8E2E928B-E1D1-4AFA-AAE7-6586F9EC8170} = {54E79D60-4472-45EF-AA93-764C706453CC}
 		{54E79D60-4472-45EF-AA93-764C706453CC} = {2C2A047E-CDC5-41E8-82C1-BCC31F941B6A}
-		{326790CA-C54A-4DC3-A574-CC054AF56CC2} = {E15E5EE4-FD58-4081-AA20-1B74C162D3E3}
-		{C522605E-EBB8-4D37-AD4B-60A096D52D34} = {E15E5EE4-FD58-4081-AA20-1B74C162D3E3}
 		{40CCCF55-F3FB-4C13-8A1D-C737699C1CAD} = {1A32473D-9490-4FF8-89E7-986DCCF4599C}
 		{94E0FC52-61C1-4498-A187-962F083F130B} = {EEF8A840-7317-4F0D-AC5F-1ED1D9197E77}
 		{29F833FD-286F-4D18-97E1-37555E510410} = {E7DF2D12-586C-4FC6-AD53-082A1B5DF85B}
@@ -211,6 +208,9 @@ Global
 		{6ECBF01A-1389-4452-BF29-C4F3FB68CADC} = {23C64FA6-FA22-4A73-8360-D81FD44E7AE1}
 		{CD33583D-EB52-4D8A-8892-18FE4840CA06} = {F4F82AA0-8F75-4A08-B420-5C5A3FC812C0}
 		{A84F8BA7-9E7D-4B49-B56B-F17B206B33B8} = {F4F82AA0-8F75-4A08-B420-5C5A3FC812C0}
+		{232C4AA8-172D-43F3-AD53-04306D70C128} = {E15E5EE4-FD58-4081-AA20-1B74C162D3E3}
+		{70A0FDC0-26E5-4480-AE9C-96C9C424AEC8} = {E15E5EE4-FD58-4081-AA20-1B74C162D3E3}
+		{55AC76A5-1297-4529-8B29-5548EA581D22} = {E15E5EE4-FD58-4081-AA20-1B74C162D3E3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8F5394A0-8892-4A22-B6A0-7BBC209FF191}
